mm-video: vdec: Keep a margin of a microsecond as tolerance
While deciding to either render or drop a given decoded frame,
if the given frame is the last frame in the vsync interval, it is
rendered. To calculate the same, sometimes due to fractional
timestamp interval differences between consecutive frames, more
frames get rendered. To compensate for such timestamp differences,
keep a tolerance of a microsecond to decide if the next predicted
frame, falls in other vsync comfortably.

Change-Id: Ice9337a2bed7b9f285ed76037078a993bc6ef446
Signed-off-by: Vikash Garodia <vgarodia@codeaurora.org>